FROMLIST: rtw88: add adaptivity support for EU/JP regulatory
From MIC Ordinance Regulating Radio Equipment article 49.20,
ETSI EN-300-328 and EN-301-893, the device should be able to
dynamically pause TX activity when energy detected on the air.

To achieve this, add Energy Detected CCA (EDCCA) support to
detect the energy in the channel. Driver will set corresponding
thresholds to the device, if the energy detected exceeds the
threshold, the TX activity will be halted immediately.

As this could lead to performance downgrade when the environment
is noisy, add a debugfs to disable this for debugging usage.

Signed-off-by: Tzu-En Huang <tehuang@realtek.com>
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>
(am from https://patchwork.kernel.org/patch/11693931/)

Conflicts:
  drivers/net/wireless/realtek/rtw88/main.h
  drivers/net/wireless/realtek/rtw88/phy.c
  drivers/net/wireless/realtek/rtw88/regd.c
  drivers/net/wireless/realtek/rtw88/rtw8822b.c
  drivers/net/wireless/realtek/rtw88/rtw8822c.c

  most of them are conflicting with RTL8723DE patch set

BUG=b:144956938
TEST=Connect to an AP is EU/JP, and see if there's less collision.

Change-Id: Iabf568a8f2f8ddf9b74a59f38df17d6a2f2818bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2274347
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
[rebase510(groeck): Substantial conflicts; re-imported from list]
Signed-off-by: Guenter Roeck <groeck@chromium.org>