FROMGIT: drm/mediatek: Refactor plane init
Add a couple of functions which enumerate the number of planes for a
component and initialize the planes for a component.

No functional changes in this patch, but it will allow us to selectively
support rotation if the component supports it.

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Signed-off-by: CK Hu <ck.hu@mediatek.com>
(cherry picked from commit 31c5558dae0513bd2bae33ea37543f584c6c35a5
 https://github.com/ckhu-mediatek/linux.git-tags.git tag mediatek-drm-next-5.5-2)

BUG=chromium:1015433
TEST=Tested on krane, set reflection/rotation verified correct image on screen

Change-Id: Ia1a8301940c1e85fd139334514bbee9d26db4442
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1904026
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2064079
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>