CHROMIUM: kernelconfig: Support running outside chroot
Since splitconfig is now stored in defconfig format, it no longer
depends on the compiler. This in turn means that we can run
kernelconfig outside the chroot without affecting the generated
configuration.

To encourage people to verify that running the command outside chroot
does not cause problems, add '-F' flag as mandatory option for the
command to run outside chroot.

BUG=chromium:1006306
TEST=Run "kernelconfig [-F] olddefconfig" outside and inside chroot and compare results

Change-Id: I312b309b4717d3a71889435c60cf0eef56c7342c
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1906112
Reviewed-by: Douglas Anderson <dianders@chromium.org>