CHROMIUM: iwl7000: pcie: fix recognition of QnJ/Hr devices
Now that the QnJ is handled in the new tables (even though it's still
hardcoded), we don't go through the mangling code below it (we jump to
the found label below it), so we need to set the cfg structs to QnJ
already.  Since all 0x2720 devices are QnJ devices, we can set it to
the QnJ struct directly (and we don't care about the strings for now).

Change-Id: I2ee894727deca824b47c1fec08c0eadf8b40814c
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 82fcfa14a3c4b99f1dc7e5448e12b7f8a100c051