UPSTREAM: vsock/vmci: register vmci_transport only when VMCI guest/host are active
To allow other transports to be loaded with vmci_transport,
we register the vmci_transport as G2H or H2G only when a VMCI guest
or host is active.

To do that, this patch adds a callback registered in the vmci driver
that will be called when the host or guest becomes active.
This callback will register the vmci_transport in the VSOCK core.

Cc: Jorgen Hansen <jhansen@vmware.com>
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit b1bba80a4376aef34de2b57bfb8834bd095703ed)

BUG=b:153004946
TEST=nested vsock works

Change-Id: I130c6d854a3f15bc76b6020fe600997a1f0b2f3d
Signed-off-by: Allen Webb <allenwebb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2406627
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>