FROMLIST: platform: chrome: sensorhub: Add median filter
Events are timestamped in EC time space, their timestamps need to be
converted in host time space.
The assumption is the time delta between when the interrupt is sent
by the EC and when it is receive by the host is a [small] constant.
This is not always true, even with hard-wired interrupt. To mitigate
worst offenders, add a median filter to weed out bigger than expected
delays.

(am from https://patchwork.kernel.org/patch/11245731/)

BUG=chromium:946178
TEST=compile

Acked-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Acked-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Acked-by: Lee Jones <lee.jones@linaro.org>
Change-Id: I6ccee259c9557bf004ad0acb1a83ac107b8926e3
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>